Just about finished the reorganization, only thing left is to reformat our autos into the new syntax which should just be copy pasting values over mostly.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mechanisms/Lift.java
@@ -48,7 +48,7 @@ public class Lift {
         transferM1.setPower(0);
         transferM2.setPower(0);
     }
-    public double getSlidePosition(){
+    public int getSlidePosition(){
         return (transferM1.getCurrentPosition() + transferM2.getCurrentPosition()) / 2;
     }
 }
